💄 enhance enemies
