+++ apps/tempest.games/src/app/globals.scss
@@ -125,7 +125,7 @@ h3 {
     --table-dark: #bbb;
   }
   background-color: var(--bg-shade-2);
-  min-height: 100vh;
+  min-height: 100dvh;
   scroll-padding-top: 4.5em;
   scroll-behavior: smooth;
   body {
