+++ apps/tempest.games/src/app/layout.module.scss
@@ -5,7 +5,7 @@ body.class {
   overflow-x: hidden;
   display: flex;
   flex-flow: column;
-  min-height: 100vh;
+  min-height: 100dvh;
   height: auto;
   > * {
     padding: 1em 0.25em;
