+++ apps/tempest.games/src/components/comic.module.scss
@@ -1,5 +1,4 @@
 .class {
-  font-family: Uruz;
   font-size: 18px;
   font-weight: 500;
   border: none;
@@ -7,7 +6,6 @@
   cursor: pointer;
   position: relative;
   background-color: transparent;
-  font-family: "Comic Sans MS";
   &:hover {
     transform: scale(1.1);
     z-index: 10000 !important;
