+++ apps/tempest.games/src/game/other-players/EnemyDomains.module.scss
@@ -11,6 +11,8 @@ div.class {
     flex-flow: row nowrap;
     height: 100%;
     gap: 10px;
+    padding: 10px;
+    padding-bottom: 15px;
     > [data-css="enemy"] {
       position: relative;
       display: flex;
@@ -20,18 +22,20 @@ div.class {
       width: 100%;
       flex: 1 0 0;
       min-width: 0;
-      gap: 10px;
       > [data-css="their-pfp"] {
-        background: #888;
-        aspect-ratio: 1 / 1;
-        max-width: 50px;
-        border-radius: 50%;
+        width: 100%;
+        display: flex;
+        flex-grow: 1;
+        justify-content: center;
+        align-items: flex-start;
+        > span {
+          height: 120%;
+        }
       }
-      > [data-css="their-hands"] {
+      > [data-css="their-stuff"] {
         position: relative;
         width: 100%;
-        max-height: 100px;
-        flex: 1 0 0;
+        flex: 6 0 0;
         max-width: 500px;
       }
     }
