+++ apps/tempest.games/src/game/other-players/TheirHands.module.scss
@@ -1,0 +1,3 @@
+div.class {
+  height: 100%;
+}
