+++ apps/tempest.games/src/game/other-players/TheirStuff.tsx
@@ -1,0 +1,7 @@
+import * as React from "react"
+
+import { TheirHands } from "./TheirHands"
+
+export const TheirStuff: React.FC<{ playerId: string }> = ({ playerId }) => {
+	return <TheirHands playerId={playerId} />
+}
