+++ apps/tempest.games/src/game/players/ProfilePicture.module.scss
@@ -1,0 +1,7 @@
+span.class {
+  --background-color: #000;
+  background-color: var(--background-color);
+  min-width: 20px;
+  aspect-ratio: 1;
+  border-radius: 50%;
+}
